Refactoring

+++ src/structs.h
@@ -1131,8 +1131,8 @@ const short LVL_IMMORT = 31;
 #define NUM_OF_DIRS  6		// number of directions in a room (nsewud) //
 #define MAGIC_NUMBER (0x06)	// Arbitrary number that won't be in a string //
 
-constexpr unsigned long long OPT_USEC = 40000u;	// 25 passes per second //
-constexpr unsigned long long PASSES_PER_SEC = 1000000u / OPT_USEC;
+constexpr long long OPT_USEC = 40000;	// 25 passes per second //
+constexpr long long PASSES_PER_SEC = 1000000 / OPT_USEC;
 
 #define RL_SEC    * PASSES_PER_SEC
 
@@ -1638,9 +1638,8 @@ struct DESCRIPTOR_DATA
 	char **history;		// History of commands, for ! mostly.  //
 	int
 	history_pos;		// Circular array position.      //
-	int bufptr;			// ptr to end of current output  //
-	int
-	bufspace;		// space left in the output buffer  //
+	size_t bufptr;			// ptr to end of current output  //
+	size_t bufspace;		// space left in the output buffer  //
 	struct txt_block *large_outbuf;	// ptr to large buffer, if we need it //
 	struct txt_q
 				input;			// q of unprocessed input     //
